packages/llm-browser/pyproject.toml
[project]
name = "llm-browser"
version = "0.2.0"
description = "Playwright browser automation with declarative YAML flows for LLM agents."
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "patchright>=1.40",
    "pyyaml>=6.0",
    "yaml-engine",
    "lxml[html_clean]>=5.0",
]

[project.optional-dependencies]
camoufox = ["camoufox>=0.4"]
nodriver = [
    "nodriver>=0.40",
]

[project.scripts]
llm-browser = "llm_browser.cli:main"

[tool.uv]
package = true

[tool.uv.sources]
yaml-engine = { path = "../yaml-engine" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "lxml-stubs>=0.5.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.11",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_browser"]

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = "yaml_engine.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "patchright.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "camoufox.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "nodriver.*"
ignore_missing_imports = true
